Grunt JS Uglify files now all vendor and 1 custom
Modified JS Uglify in Grunt to take all JS files in js/vendor under the assumption the order these files are loaded will not have a negative impact. If I find a counter-example of this behaivor, I'll need to modify this.  Only loading js/custom/javascript.js (any other files in that directory will need to be added explicitly)