load all /lib files on the genesis_setup hook
Previously, functions.php loaded all the files in /lib when the file was
executed.  Because Genesis Child Themes rely heavily on the Genesis
parent, often the child theme needs to hook in after the parent theme is
setup (on hook genesis_setup at priority 15).

By waiting to load the files in /lib until the hook genesis_setup
priority 15, we eliminate the need for extra hooking and functions in
the individual files in /lib.

See #16