Fix Primary nav margins when both menus enabled
When both primary and secondary menus are enabled, there is a 20px gap between the two menus. Bootstrap adds a margin to the bottom of the menu to create spacing between the menu and the page content, but when two menus are stacked this creates unnecessary gap between the two menus. This seems like a lot of code to add to change just one css rule, there is probably a smarter way to do this that I'm not thinking of.