Add X-UA-Compatible and viewport
 Add X-UA-Compatible and viewport here to comply with bootstraps changes:
https://github.com/twbs/bootstrap/issues/15601

This would also remove two additional files:
ie-conditional-comments.php
responsive-viewport.php


Basically unless these meta tags immediately follow the head tag (specifically x-ua-compatible), the site will break in IE-8 when compatibility mode is on. I just realized this was added to core bootstrap a few months ago. I originally came across this last year with a friend who was using Squarespace and IE-8 (which didn't have support for this at the time). This article is old but does a good job of explaining, if you're interested (http://www.validatethis.co.uk/news/fix-bad-value-x-ua-compatible-once-and-for-all/) Anyway, I think this is pretty important to add. Thoughts?